+++ frontend/src/main/resources/webapp/src/app/main/books/books-detail/books-detail.component.ts
@@ -51,9 +51,7 @@ export class BooksDetailComponent implements OnInit {
     return this.loginService.isLoggedIn()
   }
 
-  getImageSrc(base64: any) {
-    return 'data:image/png;base64,' + base64;
-  }
+
 
 
   onPublisherClick(combo: OComboComponent) {