+++ frontend/src/main/resources/webapp/src/app/main/books/cards/book-card-medium/book-card-medium.component.ts
@@ -15,9 +15,7 @@ export class BookCardMediumComponent implements OnInit {
   ngOnInit() {
   }
   
-  getImageSrc(cover: any) {
-    return 'data:image/png;base64,' + cover;
-  }
+
 
   addToMyList(book_id: number) {