+++ frontend/src/main/resources/webapp/src/app/main/publishers/cards/publisher-card-medium/publisher-card-medium.component.ts
@@ -15,9 +15,6 @@ export class PublisherCardMediumComponent implements OnInit {
   ngOnInit() {
   }
   
-  getImageSrc(logo: any) {
-    return 'data:image/png;base64,' + logo;
-  }
 
   addToMyList(publisher_id: number) {